Add test coverage for turbo-frame[disabled]
Support for the `[disabled]` attribute pre-dates this commit, but is
largely untested. There are no mentions of the attribute in the Turbo
[Handbook][] or [Reference][] documentation.

This commit adds functional test-level coverage for the existing
behavior. When an `<a>` or `<form>` element targets a `<turbo-frame
disabled>` element, Turbo skips intercepting and intervening, and
instead navigates the page.

Testing
---

In addition, to avoid conflating `[data-turbo="false"]` and
`<turbo-frame disabled>`, this commit also renames test and test fixture
mentions of "disabled" to more accurately reflect which manner of
skipping behavior is being tested.

[Handbook]: https://github.com/hotwired/turbo-site/blob/52d5e7f1dfc26ac7a6e01c6a715857ccf5ed5081/docs/handbook/03_frames.md
[Reference]: https://github.com/hotwired/turbo-site/blob/52d5e7f1dfc26ac7a6e01c6a715857ccf5ed5081/docs/reference/frames.md